Refactor library
