+++ dev/vue/main.js
@@ -2,7 +2,7 @@ import Vue from 'vue';
 import App from './App.vue';
 import './styles/main.scss';
 
-import VueDynamicForms from '../../src/main'; // Dev
+import VueDynamicForms from '../../src/index'; // Dev
 // import VueDynamicForms from '@asigloo/vue-dynamic-forms' // Prod
 
 Vue.config.productionTip = false;
