+++ dev/vue/styles/_vendors.scss
@@ -6,4 +6,4 @@
 $input-border-radius: 50px; */
 // @import '../../src/styles/themes/default.scss';
 
-@import '../../src/styles/themes/material.scss';
+@import '../../../src/styles/themes/material.scss';
